refactor: place 서비스 패키지 구조 분리 및 책임 정리 (#59)
