+++ src/main/java/com/insideout/backend/domain/place/controller/PlaceSearchController.java
@@ -3,8 +3,8 @@ package com.insideout.backend.domain.place.controller;
 import com.insideout.backend.domain.place.dto.response.PlaceNearestResponse;
 import com.insideout.backend.domain.place.dto.response.PlaceSearchItemResponse;
 import com.insideout.backend.domain.place.exception.PlaceSuccessCode;
-import com.insideout.backend.domain.place.service.PlaceSearchService;
-import com.insideout.backend.domain.place.service.PlaceSuggestService;
+import com.insideout.backend.domain.place.service.search.PlaceSearchService;
+import com.insideout.backend.domain.place.service.suggest.PlaceSuggestService;
 import com.insideout.backend.global.apiPayload.ApiResponse;
 import com.insideout.backend.global.apiPayload.code.GeneralSuccessCode;
 import io.swagger.v3.oas.annotations.Operation;
