+++ src/test/java/com/insideout/backend/domain/place/service/PlaceSearchServiceTest.java
@@ -6,6 +6,10 @@ import com.insideout.backend.domain.place.dto.response.PlaceNearestResponse;
 import com.insideout.backend.domain.place.dto.response.PlaceSearchItemResponse;
 import com.insideout.backend.domain.place.exception.PlaceErrorCode;
 import com.insideout.backend.domain.place.exception.PlaceException;
+import com.insideout.backend.domain.place.service.es.PlaceSuggestElasticsearchClient;
+import com.insideout.backend.domain.place.service.kakao.KakaoPlaceSearchClient;
+import com.insideout.backend.domain.place.service.search.PlaceSearchIndexingService;
+import com.insideout.backend.domain.place.service.search.PlaceSearchService;
 import com.insideout.backend.global.apiPayload.code.GeneralErrorCode;
 import com.insideout.backend.global.apiPayload.exception.ProjectException;
 import org.junit.jupiter.api.BeforeEach;
